test(direct-control): extract notification package tests
Move package-owned play notification view and notification dismissal coverage into focused direct-control package tests with local fake tuner fixtures.

Record the two framed notification handoffs in the support modularization workstream. The parallel agents produced boundary-clean candidates, while the DRA corrected dismissal scope, removed broad-suite notification ownership, and ran package gates in the support worktree.

Validation: bun run --cwd packages/civ7-direct-control test -- play-notification-view.test.ts notification-dismissal.test.ts; bun run --cwd packages/civ7-direct-control test -- play-notification-view.test.ts notification-dismissal.test.ts direct-control.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; git diff --check.
