+++ openspec/changes/civ7-support-direct-control-modularization/workstream/review-disposition-ledger.md
@@ -16,4 +16,5 @@
 | Current support branch has no tracked oRPC skill or control-oRPC source | `019e8af1-427e-7463-a9be-dbdeabbccfdf` | P2 | Accepted/deferred | `workstream-record.md`, `design.md`, and `tasks.md` keep oRPC planning-only until `.agents/skills/civ7-orpc-control-architecture` and/or `packages/civ7-control-orpc` authority is imported or explicitly cited from the relevant stack branches. |
 | Session/framing parallel implementation lane must not be treated as proof-complete from the agent worktree | `pendingWorktreeId local:699af531-ba6e-49e6-8f36-021dd3e27721` / disk worktree `6e69` | P1 | Accepted/repaired | The child produced only a boundary-clean candidate file; app thread read/archive tools were unavailable in this session, so DRA passively inspected the disk worktree, integrated and corrected the candidate, removed duplicate broad-suite assertions, and owns final package/OpenSpec gates. |
 | Ready-view parallel implementation lanes must not be treated as proof-complete from agent worktrees | `pendingWorktreeId local:98a2e225-27e6-4423-abff-c2df614e9547` / disk worktree `ad4e`; `pendingWorktreeId local:31e49457-5f92-4486-b597-624c44e5d18b` / disk worktree `0425` | P1 | Accepted/repaired | The children produced boundary-clean ready-unit and ready-city candidate files only; DRA passively inspected disk worktrees, integrated the candidates, removed duplicate broad-suite assertions and fake branches/helpers, and owns final package/OpenSpec gates. |
-| Restack conflict retained new broad-suite edge coverage in a file targeted for modularization | User correction during live-control submit-readiness restack | P2 | Accepted/deferred to owner workstream | `tasks.md` now has 4.2.5: notification panel-dismissal none/expired coverage must move to the notification dismissal atom, and log rewrite marker coverage must move to the proof/log support atom before source extraction. |
+| Notification parallel implementation lanes must not be treated as proof-complete from agent worktrees | `pendingWorktreeId local:e7685d6d-1350-4fe5-ab95-335382083cab` / disk worktree `27a4`; `pendingWorktreeId local:f472e74a-276d-4749-ae16-346577f0a5e9` / disk worktree `598a` | P1 | Accepted/repaired | The children produced boundary-clean candidates only; DRA passively inspected disk worktrees, corrected dismissal scope/syntax, integrated package-owned notification view/dismissal tests, removed duplicate broad-suite assertions and fake branches/helpers, and owns final package/OpenSpec gates. |
+| Restack conflict retained new broad-suite log rewrite coverage in a file targeted for modularization | User correction during live-control submit-readiness restack | P2 | Accepted/deferred to proof/log support owner workstream | `tasks.md` now has 4.2.6: remaining log rewrite marker coverage must move to the proof/log support atom before source extraction. Notification dismissal edge coverage moved to package-owned notification tests during this restack conflict. |
