+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -68,6 +68,8 @@ titles, and parent reports to close the loop.
 | `pendingWorktreeId local:699af531-ba6e-49e6-8f36-021dd3e27721` / disk worktree `6e69` | Add session/framing package tests | framed `/goal` implementation | candidate visible on disk; app thread read/archive tools unavailable in this session | yes, from disk candidate and worktree state | Accepted as boundary-clean net-new candidate only; DRA retained integration, broad-suite removal, docs, and final gates. | Created local `packages/civ7-direct-control/test/session.test.ts`; DRA added health/env-host coverage, removed duplicate broad-suite assertions, and ran package gates in the support worktree. |
 | `pendingWorktreeId local:98a2e225-27e6-4423-abff-c2df614e9547` / disk worktree `ad4e` | Add ready-unit package test | framed `/goal` implementation | candidate visible on disk; app thread read/archive tools unavailable in this session | yes, from disk candidate and worktree state | Accepted as boundary-clean net-new candidate only; DRA retained integration, broad-suite removal, docs, and final gates. | Created local `packages/civ7-direct-control/test/ready-unit-view.test.ts`; DRA kept fixture local and moved ready-unit ownership out of the broad suite. |
 | `pendingWorktreeId local:31e49457-5f92-4486-b597-624c44e5d18b` / disk worktree `0425` | Add ready-city package test | framed `/goal` implementation | candidate visible on disk; app thread read/archive tools unavailable in this session | yes, from disk candidate and worktree state | Accepted as boundary-clean net-new candidate only; DRA retained integration, broad-suite removal, docs, and final gates. | Created local `packages/civ7-direct-control/test/ready-city-view.test.ts`; DRA preserved the city identity/source guard and moved ready-city ownership out of the broad suite. |
+| `pendingWorktreeId local:e7685d6d-1350-4fe5-ab95-335382083cab` / disk worktree `27a4` | Add play notification view package test | framed `/goal` implementation | candidate visible on disk; app thread read/archive tools unavailable in this session | yes, from disk candidate and worktree state | Accepted as boundary-clean net-new candidate only; DRA retained integration, broad-suite removal, docs, and final gates. | Created local `packages/civ7-direct-control/test/play-notification-view.test.ts`; DRA kept fixture local and moved package notification-view ownership out of the broad suite. |
+| `pendingWorktreeId local:f472e74a-276d-4749-ae16-346577f0a5e9` / disk worktree `598a` | Add notification dismissal package test | framed `/goal` implementation | candidate visible on disk; app thread read/archive tools unavailable in this session | yes, from disk candidate and worktree state | Accepted as boundary-clean candidate with DRA correction; the child included a syntax bug and an extra stale-nonblocking case outside current package ownership. | DRA integrated the package-owned dismissal cases only, preserved verification semantics, removed broad-suite dismissal ownership, and left CLI bulk-dismiss/stale-nonblocking policy outside this package-test slice. |
 
 ## Parallelization Rule
 
@@ -127,6 +129,9 @@ All future agent waves must be framed before delegation:
   ownership extraction with two boundary-clean parallel net-new candidates,
   DRA-owned broad-suite removal, local package proof, and no runtime/source
   behavior claim.
+- Direct-control notification view/dismissal test slice: completed as test-only
+  package ownership extraction with two parallel net-new candidates, DRA-owned
+  correction/removal, local package proof, and no runtime/source behavior claim.
 - Review-disposition ledger: agent/reviewer findings recorded.
 - Exact dismiss-notification CLI slice: completed as test-only extraction with
   local fixture ownership and no runtime claim.
